requirements.txt
allure_python_commons==2.9.43
pytest==6.2.5
pywinauto==0.6.8
